pyproject.toml
[project]
name = "kachaka-sdk-toolkit"
version = "0.0.0"
description = "Kachaka Robot SDK unified wrapper — shared core for MCP Server and Skill"
requires-python = ">=3.10,<3.13"
license = "Apache-2.0"
authors = [{ name = "Sigma Robotics" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "kachaka-api>=3.10",
    "grpcio>=1.66",
    "mcp[cli]>=1.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-mock"]

[project.urls]
Repository = "https://github.com/sigmarobotics/kachaka-sdk-toolkit"

[project.scripts]
kachaka-setup = "kachaka_sdk_toolkit.setup_cli:main"
kachaka-mcp = "mcp_server.server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["kachaka_core*", "mcp_server*", "kachaka_sdk_toolkit*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
